% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importCSdata.R
\name{importTOA5data}
\alias{importTOA5data}
\title{imports a datafile into an R structure (type of structure depends on \code{RetOpt})}
\usage{
importTOA5data(filename, RetOpt = "all")
}
\arguments{
\item{filename}{the TOA5 file name with extension (generally .dat)}

\item{RetOpt}{Return option. As TOA5 has a header before data, you can choose to get the header (\code{RetOpt = "info"}, structure type = vector), the data (\code{RetOpt = "data"}, structure type = data.frame) or both (\code{RetOpt = "all"} (default), structure type = list).}
}
\value{
An R Structure depending on \code{RetOpt} param.
}
\description{
\code{importCSdata} reads TOA5 text file and transposes contents into an R structure
}
