% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importTOACI1.R
\name{importTOACI1data}
\alias{importTOACI1data}
\title{imports a datafile into an R structure (type of structure depends on \code{RetOpt})}
\usage{
importTOACI1data(filename, RetOpt = "data", time = "POSIXlt")
}
\arguments{
\item{filename}{the TOACI1 file name with extension (generally .dat)}

\item{RetOpt}{Return option. As TOACI1 has a header before data, you can choose to get the header (\code{RetOpt = "info"}, structure type = vector), the data (\code{RetOpt = "data"}(default), structure type = data.frame) or both (\code{RetOpt = "all"} , structure type = list).}

\item{time}{Time structure for storing TMSTAMP. May be \code{POSIXct} or \code{POSIXlt}. POSIXlt is a better solution for milisecond precision, but causes weird behavior with \code{reshape2::melt} function.}
}
\value{
An R Structure depending on \code{RetOpt} param.
}
\description{
\code{importTOACI1data} reads TOACI1 text file and transposes contents into an R structure
}
